pyproject.toml
[project]
name = "capybaradb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "numpy>=2.0.0",
    "pydantic>=2.12.3",
    "tiktoken>=0.12.0",
    "python-docx>=1.1.0",
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
    "pytest>=8.4.2",
    "pypdf>=6.1.3",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
